+++ R/package.R
@@ -3,7 +3,8 @@
 #' @name r.releases.utils-package
 #' @family help
 #' @importFrom gh gh
-#' @importFrom jsonlite read_json
+#' @importFrom jsonlite parse_json read_json
 #' @importFrom nanonext parse_url
 #' @importFrom pkgsearch cran_package
+#' @importFrom vctrs vec_rbind
 NULL